Refactor: Split predictor_types and util in multiple files
