+++ include/deadlock_checker.hpp
@@ -1,6 +1,10 @@
 # pragma once
 
+#include <cassert>
+
 #include "ord_dep_graph.hpp"
+#include "critical_section_history.hpp"
+
 
 
 // Helper struct that holds only the relevant info about a node for output 
@@ -42,11 +46,11 @@ struct std::formatter<SimpleNode> : std::formatter<std::string> {
 };
 
 struct DeadlockChecker{
-    CSHist& cs_hist;
-    DepLocToEvMapT& dep_loc_ev_map;
+    CSHist cs_hist;
+    DepLocToEvMapT dep_loc_ev_map;
     
-    DeadlockChecker(CSHist& cs_hist, DepLocToEvMapT& dep_loc_ev_map)
-        : cs_hist(cs_hist), dep_loc_ev_map(dep_loc_ev_map){}
+    DeadlockChecker(CSHist&& cs_hist, DepLocToEvMapT&& dep_loc_ev_map)
+        : cs_hist(std::move(cs_hist)), dep_loc_ev_map(std::move(dep_loc_ev_map)){}
 
     // Answers the question: Can cycle generate abstract deadlock patterns?
     bool is_abs_dlk_pattern_gen(const NodeChainT& cycle) const;
