--- include/predictor_types.hpp
@@ -1,462 +1,0 @@
-#pragma once
-
-// TODO: For CSHist, look into making the functions return const references as well
-#include <string>
-#include <vector>
-#include <format>
-#include <unordered_map>
-#include <unordered_set>
-#include <optional>
-#include <cassert>
-#include <variant>
-#include <map>
-#include <set>
-#include "vectorclock.hpp"
-#include "common_types.hpp"
-#include "util.hpp"
-
-// Format for LocksetT
-template <>
-struct std::formatter<LocksetT> : std::formatter<std::string> {
-    auto format(const LocksetT& lockset, format_context& ctx) const {
-        auto out = ctx.out();
-        for (const auto& res_id : lockset._vec)
-          std::format_to(out, "{}, ", res_id);
-        return out;
-    }
-};
-
-// TODO: TRY USING INT8_T
-// NOTE: IF ADDING MORE OR REMOVING ALWAYS BE MIDNFULL OF UPDATING from_int16
-// Event stuff
-enum class EventsT : int16_t {
-  LK = 0,
-  UK = 1,
-  RD = 2,
-  WR = 3,
-  FORK = 4,
-  JOIN = 5,
-  WAIT = 6,
-  NOTIFY = 7,
-  NOTIFYALL = 8
-};
-
-inline bool is_lock_type(EventsT ev_type){
-  return ev_type == EventsT::LK || ev_type == EventsT::UK;
-}
-
-inline bool is_access_type(EventsT ev_type){
-  return ev_type == EventsT::RD || ev_type == EventsT::WR;
-}
-
-inline bool is_th_type(EventsT ev_type){
-  return ev_type == EventsT::FORK || ev_type == EventsT::JOIN;
-}
-
-inline bool is_notif_type(EventsT ev_type){
-  return ev_type == EventsT::NOTIFY || ev_type == EventsT::NOTIFYALL;
-}
-
-
-inline bool is_cv_type(EventsT ev_type){
-  return ev_type == EventsT::WAIT || is_notif_type(ev_type);
-}
-
-inline std::optional<EventsT> from_int16(int16_t val){
-  if (val >= static_cast<int>(EventsT::LK) && 
-        val <= static_cast<int>(EventsT::NOTIFYALL)) {
-        return static_cast<EventsT>(val);
-    }
-    return std::nullopt;
-}
-
-inline EventsT unsafe_from_int16(int16_t val){
-    return static_cast<EventsT>(val);
-}
-
-// Formats EventsT
-// Needed when calling std::format
-template <>
-struct std::formatter<EventsT> : std::formatter<std::string> {
-    auto format(EventsT e, format_context& ctx) const {
-        std::string name;
-        switch (e) {
-            case EventsT::RD: name = "r"; break;
-            case EventsT::WR:  name = "w"; break;
-            case EventsT::FORK: name = "fork"; break;
-            case EventsT::JOIN: name = "join"; break;
-            case EventsT::LK: name = "acq"; break;
-            case EventsT::UK: name = "rel"; break;
-            case EventsT::WAIT: name = "wait"; break;
-            case EventsT::NOTIFY: name = "notify"; break;
-            case EventsT::NOTIFYALL: name = "notifyAll"; break;
-            default: name = "UNKNOWN"; break;
-        }
-        return formatter<std::string>::format(name, ctx);
-    }
-};
-
-struct EventInfo{
-  ThreadIdT thread_id;
-  EventsT event_type;
-  ResourceIdT target;
-  SrcLocT src_loc;
-  EventIdT line; // line in trace file 
-
-  EventInfo(){}
-  EventInfo(ThreadIdT thread_id, EventsT event_type, ResourceIdT target, SrcLocT src_loc, EventIdT line = 0)
-    : thread_id(thread_id), event_type(event_type), target(target), src_loc(src_loc), line(line){}
-};
-
-template <>
-struct std::formatter<EventInfo> : std::formatter<std::string> {
-  auto format(const EventInfo& e, format_context& ctx) const {
-      return std::format_to(ctx.out(), "Line {}: {}|{}({})|{}", e.line, e.thread_id, e.event_type, e.target, e.src_loc);
-  }
-};
-
-// An Event is defined by it's vector clock and position in the trace (so by two moments in time)
-struct Event{
-  VectorClock vc;
-  EventIdT tr_pos;
-  SrcLocT src_loc;
-  
-  Event(const VectorClock& vc, EventIdT tr_pos, SrcLocT src_loc) 
-    : vc(vc), tr_pos(tr_pos), src_loc(src_loc) {}
-  
-  Event(){}
-
-  // Compares vc and tr_pos
-  bool operator<=(const Event& other) const{
-    return vc <= other.vc && tr_pos <= other.tr_pos;
-  }
-};
-
-// Comparator between Event and VectorClock
-// The order matters! Always put vc to the right as it usually is the sync preserving closure
-struct EventComp{
-    bool operator()(const Event& ev, const VectorClock& vc) const {
-        return ev.vc < vc;
-    }
-
-    bool operator()(const VectorClock& vc, const Event& ev) const {
-        return ev.vc > vc;
-    }
-};
-
-// Comparator between Event pointer and VectorClock
-// The order matters! Always put vc to the right as it usually is the sync preserving closure
-struct EventPtrComp{
-    bool operator()(const Event* ev, const VectorClock& vc) const {
-        return ev->vc < vc;
-    }
-
-    bool operator()(const VectorClock& vc, const Event* ev) const {
-        return ev->vc > vc;
-    }
-};
-
-using EventLazyQueue = ViewLazyQueue<std::vector<Event>>;
-
-// Critical section stuff
-
-// Helper struct that hold the events for a critical section(one for lock, one for unlock)
-struct CSInfo{
-  Event lock_ev;
-  Event unlock_ev;
-
-  CSInfo(Event&& lock_ev, Event&& unlock_ev)
-    : lock_ev(std::move(lock_ev)), unlock_ev(std::move(unlock_ev)){}
-
-  CSInfo(Event&& lock_ev)
-    : lock_ev(std::move(lock_ev)){}
-  
-  CSInfo() = default;
-  
-  // Compares two critical sections using trace location
-  bool less_than_eq_tr(const CSInfo& other) const{
-    return lock_ev.tr_pos <= other.lock_ev.tr_pos && unlock_ev.tr_pos <= other.lock_ev.tr_pos;
-  }
-  
-  // Compares two critical sections using trace location
-  bool less_than_tr(const CSInfo& other) const{
-    return lock_ev.tr_pos < other.lock_ev.tr_pos && unlock_ev.tr_pos < other.lock_ev.tr_pos;
-  }
-};
-
-// Comparator between CSInfo and VectorClock'
-// The order matters! Always put vc to the right as it usually is the sync preserving closure
-struct CSInfoComp{
-    bool operator()(const CSInfo& cs, const VectorClock& vc) const {
-        return cs.lock_ev.vc < vc;
-    }
-
-    bool operator()(const VectorClock& vc, const CSInfo& cs) const {
-        return cs.lock_ev.vc > vc;
-    }
-};
-
-
-// Groups the thread ID with its queue continuously in memory
-struct ThreadCSHist {
-    ThreadIdT tid;
-    OwnedLazyQueue<std::vector<CSInfo>> queue;
-
-    ThreadCSHist(ThreadIdT tid) : tid(tid) {
-      queue.queue.reserve(10);
-    }
-};
-
-struct CSHist {
-    std::unordered_map<ResourceIdT, std::vector<ThreadCSHist>> _cs_hist;
-
-    void reset() {
-        for (auto& [res_id, th_vec] : _cs_hist) {
-            for (auto& th_hist : th_vec) {
-                th_hist.queue.reset();
-            }
-        }
-    }
-
-    CSInfo& add_lock_ev(ResourceIdT res_id, ThreadIdT tid, Event&& lock_ev) {
-        auto& th_vec = _cs_hist[res_id];
-        
-        // Linearly search for the thread
-        for (auto& th_hist : th_vec) {
-            if (th_hist.tid == tid) {
-                th_hist.queue.emplace(std::move(lock_ev));
-                return th_hist.queue.back();
-            }
-        }
-
-        // If thread hasn't used this lock yet, add it
-        th_vec.emplace_back(tid);
-        th_vec.back().queue.emplace(std::move(lock_ev));
-        return th_vec.back().queue.back();
-    }
-
-    std::optional<CSInfo*> add_unlock_ev(ResourceIdT res_id, ThreadIdT tid, Event&& unlock_ev) {
-        auto umap_it = _cs_hist.find(res_id);
-        if (umap_it == _cs_hist.end()) return {};
-
-        for (auto& th_hist : umap_it->second) {
-            if (th_hist.tid == tid) {
-                if (th_hist.queue.empty()) return {};
-                
-                CSInfo& cs = th_hist.queue.back();
-                cs.unlock_ev = std::move(unlock_ev);
-                return &cs;
-            }
-        }
-        return {};
-    }
-
-    std::optional<CSInfo*> get_back(ResourceIdT res_id, ThreadIdT tid) {
-        auto umap_it = _cs_hist.find(res_id);
-        if (umap_it == _cs_hist.end()) return {};
-
-        for (auto& th_hist : umap_it->second) {
-            if (th_hist.tid == tid) {
-                if (th_hist.queue.empty()) return {};
-                return &th_hist.queue.back();
-            }
-        }
-        return {};
-    }
-};
-
-struct StdIdMap{
-  size_t id_counter;
-  std::unordered_map<std::string, int> _map;
-
-  StdIdMap() : id_counter(0){}
-
-  void reset(){
-    id_counter = 0;
-    _map.clear();
-  }
-
-  // Returns the corresponding id for std_id from _map
-  // Updates the _map and counter if not found
-  int get(const std::string& std_id){
-    auto map_entry = _map.find(std_id);
-    int result_id;
-
-    if(map_entry == _map.end()) {
-        _map[std_id] = id_counter;
-        result_id = id_counter;
-        id_counter += 1;
-      } else {
-        result_id = map_entry->second;
-      }
-
-      return result_id;
-  }
-
-};
-
-struct LocksetEntry{
-  ThreadIdT tid;
-  size_t count;
-
-  auto operator<=>(const LocksetEntry& other) const {
-    return tid <=> other.tid;
-  }
-
-  bool operator==(const LocksetEntry& other) const {
-    return tid == other.tid;
-  }
-
-  LocksetEntry(ThreadIdT tid) : tid(tid), count(0){}
-};
-
-struct UReentrantLocksetT{
-    SortedVector<LocksetEntry> _active_locks; 
-
-    UReentrantLocksetT(){
-        _active_locks._vec.reserve(meta_info.LOCK_DEPTH); 
-    }
-
-    // Returns true if the lock is acquired for the first time
-    bool acquire(ResourceIdT lock_id){
-        LocksetEntry& entry = _active_locks.unique_insert(lock_id);
-        entry.count += 1;
-
-        return entry.count == 1;
-    }
-    
-    // Returns true if the lock is released for the last time
-    bool release(ResourceIdT lock_id){
-        bool last_time = false;
-
-        auto it = _active_locks.find(lock_id);
-        size_t& count = const_cast<size_t&>(it->count);
-        count -= 1;
-
-        if (count == 0) {
-           _active_locks._vec.erase(it);
-           last_time = true;
-        }
-
-        return last_time;
-    }
-
-    bool contains(ResourceIdT lock_id) const{
-        return _active_locks.contains(lock_id);
-    }
-
-    LocksetT to_lockset() const{
-      LocksetT lockset(_active_locks._vec.size());
-      for (int i = 0; i < lockset._vec.size(); ++i){
-        lockset._vec[i] = _active_locks._vec[i].tid;
-      }
-      return lockset;
-    }
-
-    size_t size() const{
-        return _active_locks._vec.size(); 
-    }
-
-    bool empty() const{
-        return size() <= 0;
-    }
-};
-
-struct AbsDepView {
-    ThreadIdT thread_id;
-    ResourceIdT resource_id;
-    const LocksetT& lockset;
-};
-
-struct AbsDependency{
-  ThreadIdT thread_id;
-  ResourceIdT resource_id;
-  LocksetT lockset;
-
-  AbsDependency(ThreadIdT thread_id, ResourceIdT resource_id, const LocksetT& lockset)
-    : thread_id(thread_id), resource_id(resource_id), lockset(lockset){}
-
-  // Implements all comparison operators in the default way(first compare by thread, then resource, then lockset)
-  auto operator<=>(const AbsDependency&) const = default;
-
-  // Comparison agains the view
-  auto operator<=>(const AbsDepView& rhs) const {
-        if (auto cmp = thread_id <=> rhs.thread_id; cmp != 0) return cmp;
-        if (auto cmp = resource_id <=> rhs.resource_id; cmp != 0) return cmp;
-        return lockset._vec <=> rhs.lockset._vec;
-    }
-
-  // return true if thread_ids and resource_ids differ and locksets don't intersect, false otherwise
-  bool is_valid_neigh_cand(const AbsDependency& other) const{
-    return thread_id != other.thread_id && resource_id != other.resource_id && !lockset_intersection(lockset, other.lockset);
-  }
-  
-  // return true if thread_ids and resource_ids differ and locksets don't softly intersect, false otherwise
-  bool is_valid_neigh_cand_soft(const AbsDependency& other) const{
-    return thread_id != other.thread_id && resource_id != other.resource_id && !lockset_intersection_soft(lockset, other.lockset);
-  }
-  
-  // return true if thread_ids differ and locksets don't intersect, false otherwise
-  bool is_valid_neigh_cand_opt(const AbsDependency& other) const{
-    return thread_id != other.thread_id && !lockset_intersection(lockset, other.lockset);
-  }
-
-  // TODO: Change this when separating locks and cond var
-  bool is_lock_dep() const{
-    return resource_id >= 0;
-  }
-};
-
-// The explicit std::less<> is necesssary to allow transparent lookup 
-// typedef std::map<AbsDependency, std::vector<Event>, std::less<>> NodeContainerT;
-typedef std::set<AbsDependency, std::less<>> NodeContainerT;
-typedef std::set<AbsDependency, std::less<>> AbsDepContainerT;
-
-typedef NodeContainerT::const_iterator NodeConstItT;
-typedef AbsDepContainerT::const_iterator AbsDepConstItT;
-
-typedef std::unordered_map<AbsDepConstItT, std::unordered_map<SrcLocT, std::vector<Event>>, IteratorHasher> DepLocToEvMapT;
-
-struct NodeConstItComp {
-    bool operator()(const NodeConstItT& a, const NodeConstItT& b) const {
-
-        return *a < *b;
-    }
-};
-typedef NodeConstItComp AbsDepConstItComp;
-
-typedef std::unordered_map<ResourceIdT, SortedVector<AbsDepConstItT, AbsDepConstItComp>> LockDepMapT;
-
-
-// Format for AbsDependency
-template <>
-struct std::formatter<AbsDependency> : std::formatter<std::string> {
-    auto format(const AbsDependency& dep, format_context& ctx) const {
-        return std::format_to(ctx.out(), "{}, {}, ({})", dep.thread_id, dep.resource_id, dep.lockset);
-    }
-};
-
-using SyncStatusT = std::variant<NodeConstItT, ResourceIdT>;
-
-// Custom hasher for SyncStatus
-//TODO: Can't this have collisions?
-struct SyncStatusHash {
-  std::size_t operator()(const SyncStatusT& sync_status) const {
-    if (std::holds_alternative<ResourceIdT>(sync_status)) {
-        return std::hash<ResourceIdT>{}(std::get<ResourceIdT>(sync_status));
-    }
-
-    return IteratorHasher()(std::get<NodeConstItT>(sync_status));
-  }
-};
-
-using RecentSyncStatusContT = CircularLRU<SyncStatusT, 8, SyncStatusHash>;
-
-struct ThreadInfo{
-  UReentrantLocksetT u_reen_lockset;
-  RecentSyncStatusContT recent_sync_status_cont;
-  VectorClock vec_clock;
-  bool is_asleep = false;
-
-  ThreadInfo(ThreadIdT tid) : vec_clock(tid){};
-};
