+++ include/util.hpp
@@ -3,100 +3,6 @@
 #include <vector>
 #include <string>
 #include <sstream>
-#include <cstdint>
-#include <deque>
-#include <type_traits>
-#include <concepts>
-#include <array>
-#include <fstream>
-#include <unordered_map>
-#include "common_types.hpp"
-#include "logger.hpp"
-
-struct meta{
-    ThreadIdT THREAD_COUNT = 0;
-    EventIdT EVENT_COUNT = 0;
-    ResourceIdT VAR_COUNT = 0;
-    ResourceIdT LOCK_COUNT = 0;
-    uint8_t LOCK_DEPTH = 8;
-    std::vector<uint8_t> NOTIF_THREADS;
-
-    meta(ThreadIdT THREAD_COUNT, EventIdT EVENT_COUNT, ResourceIdT VAR_COUNT, ResourceIdT LOCK_COUNT)
-        : THREAD_COUNT(THREAD_COUNT), EVENT_COUNT(EVENT_COUNT), VAR_COUNT(VAR_COUNT), LOCK_COUNT(LOCK_COUNT){}
-    
-    meta(){}
-
-    // Resized NOTIF_THREADS to THREAD_COUNT
-    void resize_notif_threads(){
-        NOTIF_THREADS.resize(THREAD_COUNT);
-    }
-
-    // Returns the size of the partial object that will be loaded from the bin/meta file
-    size_t load_sizeof() const{
-        return sizeof(*this) - sizeof(this->LOCK_DEPTH) - sizeof(NOTIF_THREADS);
-    }
-};
-
-inline meta meta_info;
-
-template <typename T, typename Compare = std::less<T>>
-struct SortedVector {
-    std::vector<T> _vec;
-    Compare _comp;
-
-    SortedVector() = default;
-    SortedVector(size_t size) : _vec(size){}
-    
-    auto operator<=>(const SortedVector& other) const {
-        return _vec <=> other._vec;
-    }
-
-    bool operator==(const SortedVector& other) const {
-        return _vec == other._vec;
-    }
-    
-    bool contains(const T& val) const {
-        auto it = std::lower_bound(_vec.begin(), _vec.end(), val, _comp);
-        return it != _vec.end() && *it == val; 
-    }
-
-    auto find(const T& val) const {
-        auto it = std::lower_bound(_vec.begin(), _vec.end(), val, _comp);
-        if (it == _vec.end() || *it != val){
-            return _vec.end();
-        }
-        return it;
-    }
-
-    void insert(const T& val){
-        auto it = std::lower_bound(_vec.begin(), _vec.end(), val, _comp);
-        _vec.insert(it, val);    
-    }
-
-    T& unique_insert(const T& val){
-        auto it = std::lower_bound(_vec.begin(), _vec.end(), val, _comp);
-        if (it == _vec.end() || *it != val){
-            it = _vec.insert(it, val);
-        }
-        return *it;
-    }
-    
-    bool erase(const T& val){
-        auto it = std::lower_bound(_vec.begin(), _vec.end(), val, _comp);
-        if (it != _vec.end() && *it == val) {
-            _vec.erase(it);
-            return true;
-        }
-        return false;
-    }
-
-    void unsafe_erase(const T& val){
-        auto it = std::lower_bound(_vec.begin(), _vec.end(), val, _comp);
-        _vec.erase(it);
-    }
-};
-
-typedef SortedVector<ResourceIdT> LocksetT;
 
 // Splits str by sep
 inline std::vector<std::string> split(const std::string& str, char sep){
@@ -112,413 +18,6 @@ inline std::vector<std::string> split(const std::string& str, char sep){
     return result;
 }
 
-// res_id should be the id of either a lock or a cond_var
-// for locks it returns the id of the associated cond var, for cond var the associated lock
-inline ResourceIdT get_ass_sync_obj(ResourceIdT res_id){
-    return -res_id;
-}
-
-inline bool is_cond_var(ResourceIdT res_id){
-    return res_id < 0;
-}
-
-// Returns true if ls1 and ls2 intersect, false otherwise
-// Currently iterates of ls2 and checks if all elements are in ls1
-inline bool lockset_intersection(const LocksetT& ls1, const LocksetT& ls2){
-    for (const auto lock : ls2._vec)
-        if (ls1.contains(lock))
-            return true;
-    return false;
-}
-
-// Returns true if ls1 and ls2 intersect, false otherwise
-// Doesn't count cond vars for intersection
-inline bool lockset_intersection_soft(const LocksetT& ls1, const LocksetT& ls2){
-    for (const auto lock : ls2._vec){
-        if (ls1.contains(lock) && !is_cond_var(lock))
-            return true;
-    }
-    return false;
-}
-
-// Insert the elements of src in dst ignoring cond vars
-// Returns true if no duplicates were found false otherwise, stopping at the first found element
-inline bool insert_lockset(const LocksetT& src, ULocksetT& dst){
-    for (const auto& res_id : src._vec){
-        if (!is_cond_var(res_id)){ // Ignore cond vars
-            auto [it, inserted] = dst.insert(res_id);
-            if (!inserted){
-                return false;
-            }
-        }
-    }
-
-    return true;
-}
-
-// Define constraint for LazyQueue that enforces the container to be either a vector or deque
-template <typename> struct is_vec_or_deq : std::false_type{};
-template <typename T, typename A> struct is_vec_or_deq<std::vector<T, A>> : std::true_type{};
-template <typename T, typename A> struct is_vec_or_deq<std::deque<T, A>> : std::true_type{};
-
-template <typename C>
-concept is_vec_or_deq_c = is_vec_or_deq<C>::value;
-
-// Queue that instead of removing front elements only moves start_elem to the right
-// For this to make sense, the internal container should be sorted(and of course T should be comparable)
-// The internal container can be either the const view of an existing container or individual one 
-template <typename ContainerT, bool is_view>
-requires is_vec_or_deq_c<ContainerT>
-struct LazyQueue {
-    // Extract the element type directly from the container
-    using T = typename ContainerT::value_type;
-
-    // StorageT is either a read only view or a self-standing container
-    using StorageT = std::conditional_t<is_view, const ContainerT*, ContainerT>;
-    
-    StorageT queue; 
-    ContainerT::const_iterator start_elem;
-
-    // Individual container
-    LazyQueue() requires (!is_view) : queue(ContainerT{}) {}
-
-    // Read only view
-    LazyQueue(const ContainerT& external) requires is_view : queue(&external) {reset();}
-
-    // FUNCTIONS FOR NON-VIEW CONTAINER
-    
-    void push(const T& x) requires (!is_view) {
-        queue.push_back(x);
-    }
-
-    template< class... Args >
-    T& emplace(Args&&... args) requires (!is_view){
-        return queue.emplace_back(std::forward<Args>(args)...);
-    }
-
-    T& back() requires (!is_view){
-        return queue.back();
-    }
-
-
-    // COMMON FUNCTIONS
-
-    const ContainerT& get() const {
-        if constexpr (is_view) {
-            return *queue;  // Dereference the pointer
-        } else {
-            return queue;   // Return the object directly
-        }
-    }
-    
-    // Returns the first element and pops
-    // If "there are no more elements" the last will be returned
-    const T* pop() {
-        if (start_elem != queue.end()) {
-            T* res = &(*start_elem);
-            ++start_elem;
-            return res;
-        }
-        return &back();
-    }
-
-    const T& back() const{
-        return get().back();
-    }
-
-    // Pops all elements that are smaller than x and returns the last element poped
-    // If inclusive is true, popping stops at the first element greater than x
-    // If all elements are greater than x optional won't have a value
-    template <typename ValT, typename CompT>
-    requires std::predicate<CompT, const T&, const ValT&>
-    std::pair<std::optional<const T*>, bool> pop_until(const ValT& x, CompT comp, bool inclusive) {
-        const ContainerT& q = get();
-        auto tmp = start_elem;
-
-        if (inclusive){
-            start_elem = std::upper_bound(start_elem, q.cend(), x, comp);
-        }
-        else{
-            start_elem = std::lower_bound(start_elem, q.cend(), x, comp);
-        }
-
-        if (start_elem == q.begin())
-            return {std::nullopt, false};
-        
-        return {&(*std::prev(start_elem)), tmp == start_elem};
-    }
-
-    bool empty() const{
-        return start_elem == get().cend();
-    }
-
-    size_t size() const{
-        return get().size();
-    }
-
-    // Sets start_elem to the beginning of the internal container
-    void reset() {
-        start_elem = get().cbegin();
-    }
-    
-};
-
-// Aliases for LazyQueue
-template <typename ContainerT>
-using OwnedLazyQueue = LazyQueue<ContainerT, false>;
-
-template <typename ContainerT>
-using ViewLazyQueue = LazyQueue<ContainerT, true>;
-
-// Fixed size, double ended circular linked list
-template <typename T, size_t Capacity>
-struct CircularList {
-    struct Node {
-        std::optional<T> data;
-        int prev = -1;
-        int next = -1;
-    };
-
-    template <bool IsConst>
-    struct IteratorBase {
-        using NodePtrT = std::conditional_t<IsConst, const Node*, Node*>;
-        NodePtrT _ptr;
-        NodePtrT _base_ptr; // Used to transform indexes to pointers
-
-        using PtrT = std::conditional_t<IsConst, const Node*, Node*>;
-        using RefT = std::conditional_t<IsConst, const Node&, Node&>;
-        
-        IteratorBase(Node* base_ptr, Node* ptr = nullptr) : _base_ptr(base_ptr), _ptr(ptr){}
-        IteratorBase(Node* base_ptr, int idx = -1) : _base_ptr(base_ptr){
-            _ptr = ind_to_ptr(idx);
-        }
-
-        // Conversion from iterator to const_iterator
-        template <bool WasConst, typename = std::enable_if_t<IsConst && !WasConst>>
-        IteratorBase(const IteratorBase<WasConst>& other) : _ptr(other._ptr) {}
-
-        NodePtrT ind_to_ptr(int index){
-            NodePtrT res = index == -1 ? nullptr : _base_ptr + index;
-            return res;
-        }
-
-        int to_ind(){
-            return _ptr - _base_ptr;
-        }
-        
-        RefT& operator*() {return *_ptr;}
-        PtrT operator->() const {return _ptr;}
-        IteratorBase& operator++() {_ptr = ind_to_ptr(_ptr->next); return *this;}
-        bool operator!=(const IteratorBase& other) const {return _ptr != other._ptr;}
-        bool operator==(const IteratorBase& other) const {return _ptr == other._ptr;}
-    };
-
-    using iterator = IteratorBase<false>;
-    using const_iterator = IteratorBase<true>;
-    
-    std::array<Node, Capacity> nodes;
-    
-    int head = -1;      // Points to the oldest element
-    int tail = -1;      // Points to the newest element
-    int free_head = 0;  // Points to the first available empty slot
-    size_t size = 0;
-
-    CircularList() {
-        // Initialize the free list to chain all empty slots together
-        for (int i = 0; i < Capacity - 1; ++i) {
-            nodes[i].next = i + 1;
-        }
-        nodes[Capacity - 1].next = -1;
-    }
-
-    // Insert at the end. If full, overwrites the oldest element
-    // Returns the erased element's value and the index of the inserted element
-    std::pair<std::optional<T>, int> push(const T& value) {
-        int new_node_idx;
-        std::optional<T> erased_elem;
-
-        if (full()) {
-            erased_elem = nodes[head].data;
-            new_node_idx = head;
-            
-            head = nodes[head].next;
-            nodes[head].prev = -1;
-            
-            size--;
-        } else {
-            // Pop an empty slot from the free list
-            new_node_idx = free_head;
-            free_head = nodes[free_head].next;
-        }
-
-        // Setup the new node's data and links
-        nodes[new_node_idx].data = value;
-        nodes[new_node_idx].next = -1;
-        nodes[new_node_idx].prev = tail;
-
-        if (tail != -1) {
-            nodes[tail].next = new_node_idx;
-        }
-        tail = new_node_idx;
-
-        // If this is the first element, it's also the head
-        if (head == -1) {
-            head = new_node_idx;
-        }
-
-        size++;
-        return {erased_elem, new_node_idx};
-    }
-
-    // Remove from anywhere using index
-    // Doesn't check for a valid index
-    T unsafe_erase(int target_idx) {
-        Node& target = nodes[target_idx];
-
-        // Unlink from the active list
-        if (target.prev != -1) {
-            nodes[target.prev].next = target.next;
-        } else {
-            head = target.next;
-        }
-
-        if (target.next != -1) {
-            nodes[target.next].prev = target.prev;
-        } else {
-            tail = target.prev;
-        }
-
-        T result = std::move(target.data.value());
-
-        // Clear the data and push the slot back onto the free list
-        target.data.reset();
-        target.prev = -1;
-        target.next = free_head;
-        free_head = target_idx;
-
-        size--;
-        
-        // Reset list if it becomes empty
-        if (empty()) {
-            head = -1;
-            tail = -1;
-        }
-
-        return result;
-    }
-
-    // Remove from anywhere using iterator
-    // Doesn't check for a valid iterator
-    T unsafe_erase(iterator target_it) {
-        int target_idx = target_it.to_ind();
-        return unsafe_erase(target_idx);
-    }
-
-    // Makes the element at target_idx be the newest
-    // Assumes target_idx is valid
-    void unsafe_update(int target_idx){
-        // Updating the last element doesn't do anything
-        if (target_idx == tail){
-            return;
-        }
-
-        Node& target = nodes[target_idx];
-
-        // Set target's neighbours pointers
-        if (target_idx == head){
-            head = target.next;
-        }
-        else{
-            nodes[target.prev].next = target.next;
-        }
-
-        nodes[target.next].prev = target.prev;
-        
-        // Set target pointers
-        target.next = -1;
-        target.prev = tail;
-
-        // Set the tail
-        nodes[tail].next = target_idx;
-        tail = target_idx;
-    }
-
-    template <typename Self>
-    auto begin(this Self&& self) {
-        constexpr bool is_const = std::is_const_v<std::remove_reference_t<Self>>;
-        return IteratorBase<is_const>(self.nodes.data(), self.head); 
-    }
-
-    template <typename Self>
-    auto end(this Self&& self) {
-        constexpr bool is_const = std::is_const_v<std::remove_reference_t<Self>>;
-        return IteratorBase<is_const>(self.nodes.data(), nullptr); 
-    }
-
-    template <typename Self>
-    auto last(this Self&& self) {
-        constexpr bool is_const = std::is_const_v<std::remove_reference_t<Self>>;
-        return IteratorBase<is_const>(self.nodes.data(), self.tail); 
-    }
-
-    bool full() const{
-        return size == Capacity;
-    }
-
-    bool empty() const{
-        return size == 0;
-    }
-};
-
-// Fixed size, circular LRU cache
-template <typename T, size_t Capacity, typename Hash>
-struct CircularLRU{
-    using container_t = CircularList<T, Capacity>;
-    using iterator = container_t::iterator;
-    using const_iterator = container_t::const_iterator;
-
-    container_t container;
-    std::unordered_map<T, int, Hash> idx_umap;
-
-    void push(const T& value){
-        // Update if element already present
-        auto value_idx_it = idx_umap.find(value);
-        if (value_idx_it != idx_umap.end()){
-            container.unsafe_update(value_idx_it->second);
-            return;
-        }
-
-        // Push new
-        auto [old_val_opt, new_val_ind] = container.push(value);
-        idx_umap[value] = new_val_ind;
-        
-        // Erase old val from map
-        if (old_val_opt.has_value()){
-            idx_umap.erase(old_val_opt.value());
-        }
-    }
-
-    // Updates the value pointed by target_it to be new_value
-    // Assumes target_it is a valid iterator
-    void unsafe_set(iterator target_it, const T& new_value){
-        int target_idx = target_it.to_ind();
-        unsafe_set(target_idx, new_value);
-    }
-
-    // Updates the value pointed by target_idx to be new_value
-    // Assumes target_idx is a valid index
-    void unsafe_set(size_t target_idx, const T& new_value){
-        auto& target = container.nodes[target_idx];
-
-        // Set the new value
-        T old_value = std::move(target.data.value());
-        target.data = new_value;
-
-        // Update the index map
-        idx_umap.erase(old_value);
-        idx_umap.insert({new_value, target_idx}); 
-    }
-};
 
 // Generic struct used for pointer comparison
 // Note: Nullptr is treated as infinity
@@ -534,6 +33,21 @@ struct PtrLess {
         
         return *a < *b;
     }
+};
+
+// Formatter for pointers, just formats the object it point to
+// Ignores void and char ptrs, makes sure the underlying type T is formatable
+template <typename T>
+requires (!std::same_as<std::remove_cv_t<T>, void>) &&  
+         (!std::same_as<std::remove_cv_t<T>, char>) &&
+         std::formattable<T, char>
+struct std::formatter<T*> :  std::formatter<std::string>{
+  
+  auto format(const T* ptr, auto& ctx) const {
+    if (!ptr) 
+        return std::format_to(ctx.out(), "nullptr");
+    return std::format_to(ctx.out(), "{}", *ptr);
+  }
 };
